core: Fix crash with newer vulkan-hpp
Newer vulkan-hpp headers changed the
vk::PhysicalDeviceProperties::device_name variable to use an array
wrapper class instead of being a pure character array. The address of
this object is not a string anymore, so passing it to Log::info() causes
a crash.

The array wrapper class supports a type conversion cast to a pointer, so
we can can just cast this object to 'char const*' to get a pointer to
the underlying character string. This fixes the crash with newer
vulkan-hpp versions, while still being correct for older versions.

Co-Authored-By: Matt Scheirer <matt.scheirer@gmail.com>