src/meson.build: revert to older form to find vx.xml
The last minute upstream change #76 to compute the sys_root inside of
join_paths broke cross compilation. This is because the behaviour of
join_paths is to override the built up path when an absolute path is
in place. As a result we see:

  sysroot -> /home/alex/lsrc/tests/buildroot.git/builds/arm64/host/aarch64-buildroot-linux-gnu/sysroot
  prefix -> /usr

then the final path becomes:

  /usr/share/vulkan/registry/vk.xml

I think the longer term solution would be to propose the registry path
is exported by vulkan itself so the vkmark meson.build doesn't have to
play these games.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>